pyproject.toml
[project]
name = "geminichatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "google-generativeai>=0.8.5",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.1.0",
]
